requirements.txt
streamlit==1.38.0
openai==1.45.0
python-dotenv==1.0.1
chromadb==0.5.5
pypdf==4.3.1
httpx==0.27.0
